+++ src/main/java/io/cem/modules/cem/entity/RecordWebPageEntity.java
@@ -1,7 +1,6 @@
 package io.cem.modules.cem.entity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -38,21 +37,21 @@ public class RecordWebPageEntity implements Serializable {
 	//记录时间
 	private String recordTime;
 	//DNS时延
-	private BigDecimal dnsDelay;
+	private Double dnsDelay;
 	//连接时延
-	private BigDecimal connDelay;
+	private Double connDelay;
 	//首字节时延
-	private BigDecimal headbyteDelay;
+	private Double headbyteDelay;
 	//页面文件时延
-	private BigDecimal pageFileDelay;
+	private Double pageFileDelay;
 	//重定向时延
-	private BigDecimal redirectDelay;
+	private Double redirectDelay;
 	//首屏时延
-	private BigDecimal aboveFoldDelay;
+	private Double aboveFoldDelay;
 	//页面元素时延
-	private BigDecimal pageElementDelay;
+	private Double pageElementDelay;
 	//下载速率
-	private BigDecimal downloadRate;
+	private Double downloadRate;
 	//备注
 	private String remark;
 
@@ -287,97 +286,97 @@ public class RecordWebPageEntity implements Serializable {
 	/**
 	 * 设置：DNS时延
 	 */
-	public void setDnsDelay(BigDecimal dnsDelay) {
+	public void setDnsDelay(Double dnsDelay) {
 		this.dnsDelay = dnsDelay;
 	}
 	/**
 	 * 获取：DNS时延
 	 */
-	public BigDecimal getDnsDelay() {
+	public Double getDnsDelay() {
 		return dnsDelay;
 	}
 	/**
 	 * 设置：连接时延
 	 */
-	public void setConnDelay(BigDecimal connDelay) {
+	public void setConnDelay(Double connDelay) {
 		this.connDelay = connDelay;
 	}
 	/**
 	 * 获取：连接时延
 	 */
-	public BigDecimal getConnDelay() {
+	public Double getConnDelay() {
 		return connDelay;
 	}
 	/**
 	 * 设置：首字节时延
 	 */
-	public void setHeadbyteDelay(BigDecimal headbyteDelay) {
+	public void setHeadbyteDelay(Double headbyteDelay) {
 		this.headbyteDelay = headbyteDelay;
 	}
 	/**
 	 * 获取：首字节时延
 	 */
-	public BigDecimal getHeadbyteDelay() {
+	public Double getHeadbyteDelay() {
 		return headbyteDelay;
 	}
 	/**
 	 * 设置：页面文件时延
 	 */
-	public void setPageFileDelay(BigDecimal pageFileDelay) {
+	public void setPageFileDelay(Double pageFileDelay) {
 		this.pageFileDelay = pageFileDelay;
 	}
 	/**
 	 * 获取：页面文件时延
 	 */
-	public BigDecimal getPageFileDelay() {
+	public Double getPageFileDelay() {
 		return pageFileDelay;
 	}
 	/**
 	 * 设置：重定向时延
 	 */
-	public void setRedirectDelay(BigDecimal redirectDelay) {
+	public void setRedirectDelay(Double redirectDelay) {
 		this.redirectDelay = redirectDelay;
 	}
 	/**
 	 * 获取：重定向时延
 	 */
-	public BigDecimal getRedirectDelay() {
+	public Double getRedirectDelay() {
 		return redirectDelay;
 	}
 	/**
 	 * 设置：首屏时延
 	 */
-	public void setAboveFoldDelay(BigDecimal aboveFoldDelay) {
+	public void setAboveFoldDelay(Double aboveFoldDelay) {
 		this.aboveFoldDelay = aboveFoldDelay;
 	}
 	/**
 	 * 获取：首屏时延
 	 */
-	public BigDecimal getAboveFoldDelay() {
+	public Double getAboveFoldDelay() {
 		return aboveFoldDelay;
 	}
 	/**
 	 * 设置：页面元素时延
 	 */
-	public void setPageElementDelay(BigDecimal pageElementDelay) {
+	public void setPageElementDelay(Double pageElementDelay) {
 		this.pageElementDelay = pageElementDelay;
 	}
 	/**
 	 * 获取：页面元素时延
 	 */
-	public BigDecimal getPageElementDelay() {
+	public Double getPageElementDelay() {
 		return pageElementDelay;
 	}
 	/**
 	 * 设置：下载速率
 	 */
-	public void setDownloadRate(BigDecimal downloadRate) {
+	public void setDownloadRate(Double downloadRate) {
 		this.downloadRate = downloadRate;
 	}
 	/**
 	 * 获取：下载速率
 	 */
-	public BigDecimal getDownloadRate() {
+	public Double getDownloadRate() {
 		return downloadRate;
 	}
 	/**
